Java, SQL libraries updated

+++ datasets/java/construction_permits/src/corgis/construction_permits/domain/Valuations.java
@@ -19,7 +19,7 @@ public class Valuations {
     private Integer twoUnits;
     private Integer fiveplusUnits;
     private Integer threeFourUnits;
-    private Long oneUnit;
+    private Integer oneUnit;
     
     
     /*
@@ -52,7 +52,7 @@ public class Valuations {
     /*
      * @return 
      */
-    public Long getOneUnit() {
+    public Integer getOneUnit() {
         return this.oneUnit;
     }
     
@@ -77,7 +77,7 @@ public class Valuations {
             this.twoUnits = ((Number)json_data.get("2 units")).intValue();// 5+ units
             this.fiveplusUnits = ((Number)json_data.get("5+ units")).intValue();// 3-4 units
             this.threeFourUnits = ((Number)json_data.get("3-4 units")).intValue();// 1 unit
-            this.oneUnit = ((Number)json_data.get("1 unit")).longValue();
+            this.oneUnit = ((Number)json_data.get("1 unit")).intValue();
         } catch (NullPointerException e) {
     		System.err.println("Could not convert the response to a Valuations; a field was missing.");
     		e.printStackTrace();
